Do not calculate VCI if NDVI data are not in database.

+++ src/drought.py
@@ -32,30 +32,34 @@ def _movingAverage(data, n):
 
 def calcVCI(model, table="ndvi.modis"):
     """Calculate Vegetation Condition Index."""
+    log = logging.getLogger(__name__)
     sdate = date(model.startyear, model.startmonth, model.startday).strftime("%Y-%m-%d")
     edate = date(model.endyear, model.endmonth, model.endday).strftime("%Y-%m-%d")
-    db = dbio.connect(model.dbname)
-    cur = db.cursor()
-    sql = "create table ndvi_max as (select st_union(rast, 'MAX') as rast from {0})".format(table)
-    cur.execute(sql)
-    sql = "create table ndvi_min as (select st_union(rast, 'MIN') as rast from {0})".format(table)
-    cur.execute(sql)
-    sql = "create table ndvi_max_min as (select st_mapalgebra(max.rast, 1, min.rast, 1, '[rast1]-[rast2]') as rast from ndvi_max as max, ndvi_min as min)"
-    Cur.execute(sql)
-    db.commit()
-    sql = "create table f1 as (select fdate, st_mapalgebra(f.rast, 1, min.rast, 1, '[rast1]-[rast2]') as rast from {0} as f, ndvi_min as min where fdate>=date'{1}' and fdate<=date'{2}' group by fdate,f.rast,min.rast)".format(table, sdate, edate)
-    cur.execute(sql)
-    db.commit()
-    sql = "create table {0}.vci as (select fdate, st_mapalgebra(f1.rast, 1, mm.rast, 1, '[rast1]/[rast2]') as rast from f1, ndvi_max_min as mm group by fdate,f1.rast,mm.rast)".format(model.name)
-    cur.execute(sql)
-    db.commit()
-    cur.execute("drop table ndvi_max")
-    cur.execute("drop table ndvi_min")
-    cur.execute("drop table ndvi_max_min")
-    cur.execute("drop table f1")
-    db.commit()
-    cur.close()
-    db.close()
+    if dbio.tableExists(model.dbname, table.split(".")[0], table.split(".")[1]):
+        db = dbio.connect(model.dbname)
+        cur = db.cursor()
+        sql = "create table ndvi_max as (select st_union(rast, 'MAX') as rast from {0})".format(table)
+        cur.execute(sql)
+        sql = "create table ndvi_min as (select st_union(rast, 'MIN') as rast from {0})".format(table)
+        cur.execute(sql)
+        sql = "create table ndvi_max_min as (select st_mapalgebra(max.rast, 1, min.rast, 1, '[rast1]-[rast2]') as rast from ndvi_max as max, ndvi_min as min)"
+        Cur.execute(sql)
+        db.commit()
+        sql = "create table f1 as (select fdate, st_mapalgebra(f.rast, 1, min.rast, 1, '[rast1]-[rast2]') as rast from {0} as f, ndvi_min as min where fdate>=date'{1}' and fdate<=date'{2}' group by fdate,f.rast,min.rast)".format(table, sdate, edate)
+        cur.execute(sql)
+        db.commit()
+        sql = "create table {0}.vci as (select fdate, st_mapalgebra(f1.rast, 1, mm.rast, 1, '[rast1]/[rast2]') as rast from f1, ndvi_max_min as mm group by fdate,f1.rast,mm.rast)".format(model.name)
+        cur.execute(sql)
+        db.commit()
+        cur.execute("drop table ndvi_max")
+        cur.execute("drop table ndvi_min")
+        cur.execute("drop table ndvi_max_min")
+        cur.execute("drop table f1")
+        db.commit()
+        cur.close()
+        db.close()
+    else:
+        log.warning("No NDVI data were found in database. Cannot calculate VCI!")
     return None
